refactor agent file
