+++ ps_genai_agents/agents/__init__.py
@@ -1,11 +1,4 @@
-from .agents import (
-    create_simple_text2cypher_agent,
-    create_text2cypher_with_visualization_agent,
-    create_visualization_agent,
-)
+from .text2cypher import create_text2cypher_agent
+from .visualization import create_visualization_agent
 
-__all__ = [
-    "create_simple_text2cypher_agent",
-    "create_text2cypher_with_visualization_agent",
-    "create_visualization_agent",
-]
+__all__ = ["create_text2cypher_agent", "create_visualization_agent"]
