+++ ps_genai_agents/components/visualize/correct_details/node.py
