+++ ps_genai_agents/components/visualize/generate_details/__init__.py
