+++ ps_genai_agents/components/visualize/generate_details/models.py
