+++ ps_genai_agents/components/visualize/generate_details/node.py
