+++ ps_genai_agents/components/visualize/generate_details/prompts.py
