+++ ps_genai_agents/components/visualize/validate_details/node.py
